Update generateRasa.ts
The generateRasa.ts file has been updated to include a new shell command to list the contents of the rasaPath directory after copying the templates. Additionally, another shell command has been added to list the contents of the basePath directory in order to get the base models. Finally, a shell command has been added to copy the generated model to a remote server.