Update generateRasa.ts to fix file path in shell command
The file generateRasa.ts was updated to fix a file path issue in a shell command. The command `cd ${basePath}/models/ && cp ${Customer.id}-${Customer.version}.tar.gz ${process.cwd()}/models` was modified to correctly copy the file to the models directory. This change ensures that the file is copied to the intended location and improves the overall functionality of the code.