Update docker-compose.yml and add STT service
Added a new service called stt to the docker-compose.yml file. This service is responsible for speech-to-text functionality and requires GPU capabilities. It has been configured with the necessary environment variables and volumes.

Also, added two new files: services/stt/Dockerfile and services/stt/nats_stt.py. The Dockerfile installs the required dependencies for the STT service, while nats_stt.py contains the code for processing speech-to-text requests using NATS messaging.

In addition, made some changes to the package-lock.json and package.json files in the services/webapp directory. Added the @types/howler dependency to both files.

These changes enhance the functionality of the project by adding the STT service and updating the dependencies for the webapp service.