Update .gitignore and docker-compose.yml
The .gitignore file was modified to include the templates/rasa/domain.yml file. The docker-compose.yml file was modified to change the volume paths for the /root/.cache/huggingface and /root/.ssh directories to ~/.cache and ~/.ssh respectively. This change ensures that the correct directories are mounted in the container.

BEGIN SAMPLE COMMIT MESSAGE
Update .gitignore and docker-compose.yml

The .gitignore file was modified to include the templates/rasa/domain.yml file. The docker-compose.yml file was modified to change the volume paths for the /root/.cache/huggingface and /root/.ssh directories to ~/.cache and ~/.ssh respectively. This change ensures that the correct directories are mounted in the container.
END SAMPLE COMMIT MESSAGE